Split code specific for certain editors into separate utils.
